+++ erpsystem/src/main/java/com/erp/erpsystem/db/Liability.java
@@ -1,5 +1,6 @@
 package com.erp.erpsystem.db;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 import jakarta.persistence.Column;
@@ -23,19 +24,19 @@ public class Liability {
     private String name; // 부채 이름
 
     @Column(nullable = false)
-    private Double currentValue; // 부채의 현재 가치
+    private BigDecimal currentValue; // 부채의 현재 가치
 
     @Column(length = 100)
     private String type; // 부채 유형 (유동/비유동 등)
 
     @Column(nullable = false)
-    private Double originValue; // 부채의 원래 가치
+    private BigDecimal originValue; // 부채의 원래 가치
 
     @Column(nullable = false)
     private LocalDateTime date; // 부채 발생 날짜
 
     @Column(nullable = false)
-    private Double interestRate; // 이자율
+    private BigDecimal interestRate; // 이자율
 
     @Column(nullable = false, length = 100)
     private String interestPeriod; // 이자 주기
